file download handler moved to OnlyOfficeFileProviderServlet
